feat: Revised version of Poap contract
This is a new version of Poap smart contract with the following changes:

1) We use zeppelinOS, and thus support upgrades for our contract

2) We assign each token to an event, thus we don't need the tokenURI
for each, we can generate one on the fly

3) All roles and permissions were revisited. Only two roles: admin and
event minters.

4) SmartContract paused state revisited. All "write" operations are not
behind the 'whenNotPaused'. Only admin can pause/unpause the contract

5) Mint is done by event. Event organizers can now have specific permissions
to mint token for their event only.

6) Batch minting: Is a common operation among organizers to assign a
batch of tokens to a batch of addresses (atendees). A new method for
this use case was added.

Next steps: Test and document the smart contract